Create drop target overlay
Create an overlay to be shown on top of the important list and act
as the actual drop target. This will allow the list elements themselves
to be moved freely to make room for the dropped item, without affecting
the drop targets as used for the drag and drop operation.